+++ builder/azure/arm/builder_acc_test.go
@@ -25,7 +25,7 @@ import (
 	"fmt"
 	"os"
 
-	builderT "github.com/hashicorp/packer/helper/builder/testing"
+	builderT "github.com/hashicorp/packer/packer-plugin-sdk/builder/testing"
 )
 
 const DeviceLoginAcceptanceTest = "DEVICELOGIN_TEST"
