+++ builder/osc/bsu/builder_acc_test.go
@@ -4,7 +4,7 @@ package bsu
 import (
 	"testing"
 
-	builderT "github.com/hashicorp/packer/helper/builder/testing"
+	builderT "github.com/hashicorp/packer/packer-plugin-sdk/builder/testing"
 )
 
 func TestBuilderAcc_basic(t *testing.T) {
