+++ builder/virtualbox/iso/builder_acc_test.go
@@ -5,7 +5,7 @@ import (
 	"path/filepath"
 	"testing"
 
-	builderT "github.com/hashicorp/packer/helper/builder/testing"
+	builderT "github.com/hashicorp/packer/packer-plugin-sdk/builder/testing"
 )
 
 func TestBuilderAcc_basic(t *testing.T) {
