+++ builder/yandex/builder_acc_test.go
@@ -7,7 +7,7 @@ import (
 
 	"github.com/go-resty/resty/v2"
 
-	builderT "github.com/hashicorp/packer/helper/builder/testing"
+	builderT "github.com/hashicorp/packer/packer-plugin-sdk/builder/testing"
 )
 
 const InstanceMetadataAddr = "169.254.169.254"
